git-svn-id: file:///home/danlipsa/dan-svn-repo/src@2556 d596a3c1-de11-4375-84b4-10211c464620

+++ ViewSettings.h
@@ -219,11 +219,11 @@ public:
     G3D::Matrix3 GetRotationForAxesOrder (const Foam& foam) const;
 
     // scale
-    double GetScaleRatio () const
+    float GetScaleRatio () const
     {
 	return m_scaleRatio;
     }
-    void SetScaleRatio (double scaleRatio)
+    void SetScaleRatio (float scaleRatio)
     {
 	m_scaleRatio = scaleRatio;
     }
@@ -661,7 +661,7 @@ private:
     G3D::Matrix3 m_rotationFocus;
     G3D::Vector3 m_rotationCenter;
     RotationCenterType m_rotationCenterType;
-    double m_scaleRatio;
+    float m_scaleRatio;
     double m_gridScaleRatio;
     double m_contextScaleRatio;
     G3D::Vector3 m_translation;
